Factor article image comps out of generic image module
